#186 rename SigningJob interface name
