+++ src/main/java/digital/slovensko/autogram/ui/gui/PDFAComplianceDialogController.java
@@ -1,6 +1,5 @@
 package digital.slovensko.autogram.ui.gui;
 
-import digital.slovensko.autogram.core.ISigningJob;
 import digital.slovensko.autogram.core.SigningJob;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -8,7 +7,7 @@ import javafx.scene.Node;
 import javafx.scene.control.Button;
 
 public class PDFAComplianceDialogController implements SuppressedFocusController {
-    private final ISigningJob job;
+    private final SigningJob job;
     private final GUI gui;
 
     @FXML
@@ -18,7 +17,7 @@ public class PDFAComplianceDialogController implements SuppressedFocusController
     @FXML
     Node mainBox;
 
-    public PDFAComplianceDialogController(ISigningJob job, GUI gui) {
+    public PDFAComplianceDialogController(SigningJob job, GUI gui) {
         this.job = job;
         this.gui = gui;
     }
